Remove invalid_texture_contenttype test SW3D-727
This test was incorrectly validating the contenttype attribute of the model texture2d element.  It was checking to see if the contenttype value was found in the [ContentTypes.xml] file.  The correct way to validate this value is to validate against the materials spec's xsd schema which will check to ensure that the value is either "image/jpeg" or "image/png".  This schema validation will be covered once we get to the materials spec.

<m:texture2d id="2" path="/3D/Textures/texture.png" contenttype="image/ping"/>